🌟 [Major]: Introducing classes and removing `base64` encoding (#66)
## Description

This pull request introduces several new classes, removes redundant
functions, and refactors existing code to improve the management of
contexts and their settings. The most significant changes include the
addition of the `Context` and `ContextInfo` classes, the removal of
Base64-related functions, and updates to the handling of context
information.

### Removal of Base64 encoding

Base64 encoding of names has been removed from the module for now. We
want to make it more flexible for module developers, so this can be
decided on the implementation of `Context` in your modules. However take
into consideration that `Secret` names (from the
`Microsoft.PowerShell.SecretManagement` module) cannot have some special
characters, like `[]` for now.

* Removed Base64-related functions: `ConvertFrom-Base64`,
`ConvertTo-Base64`, and `Test-Base64` from the
`src/functions/private/Base64` directory.
[[1]](diffhunk://#diff-8a5b41d97e0042dd2fcbb1887e4e61033bdc9fff6177f41f11c5182107612fb7L1-L24)
[[2]](diffhunk://#diff-6aa9bf88c2e78de4b6c7634a0117d6e84713a537f6e2e210238b3033ace814ebL1-L23)
[[3]](diffhunk://#diff-6759b1a6909f449203ed08424b482ec30b384370d99327c86a29df96c206d384L1-L26)

### New Classes:

*
[`src/classes/public/Context.ps1`](diffhunk://#diff-7a51bea35b52332a9254d03ce9bd0460db366f01865f76dffb47e34acd759d3dR1-R29):
Added a new `Context` class that includes methods for creating context
objects from different sources and a method to return the context ID.
*
[`src/classes/public/ContextInfo.ps1`](diffhunk://#diff-c0d9cec44494337fa03fe17fed9e3a5ebfb8232d6049aacf61042b657bfcec05R1-R19):
Added a new `ContextInfo` class that includes methods for creating
context info objects from different sources.

### Remove `ContextSetting` functions:

* Removing the `ContextSetting` functions so that the implementer can
create their own implementation. We want to push using `Get-Context` to
get the object loaded to memory from disk, modify the object in memory
(allows to modify one or more settings) and then use `Set-Context` to
write the context object back to disk when all is done. Mantra: Work
with objects.
* Removed context setting functions: `Get-ContextSetting`,
`Remove-ContextSetting`, and `Set-ContextSetting` from the
`src/functions/public/ContextSetting` directory.
[[1]](diffhunk://#diff-a7e7c3be32ea7439a398becda0d639720280a5590a59103b9dbd2f9323340927L1-L53)
[[2]](diffhunk://#diff-a538c8e1de3b76b734d62166601cdb77c155c66918bfc9ed180671425c3af1edL1-L74)
[[3]](diffhunk://#diff-738db29a4c3ef4bad20bd00b8c218ca326ab7094c4d48fab490df89e03d0b9e1L1-L72)

### Refactoring:

* Made `Get-ContextInfo` function public
`src/functions/public/Get-ContextInfo.ps1`.
* Drive consistency on how to identify a `Context`, which are now
identified by their `ID` throughout the functions.
*
[`src/completers.ps1`](diffhunk://#diff-792b150421507ebd45a85f5eedf88b6fc3ead29405ceff31779746e5cd5bf7c7L1-R9):
Updated the argument completer to use `ID` instead of `Name` for context
identification.
*
[`src/functions/public/Get-Context.ps1`](diffhunk://#diff-0d7db8bc789bd7b13874d319644d59a93cd442baa4da9f9a20854434d7a2b41aL22-R22):
Modified the `Get-Context` function to use `ID` instead of `Name` for
context identification and added indentation to debug messages.
[[1]](diffhunk://#diff-0d7db8bc789bd7b13874d319644d59a93cd442baa4da9f9a20854434d7a2b41aL22-R22)
[[2]](diffhunk://#diff-0d7db8bc789bd7b13874d319644d59a93cd442baa4da9f9a20854434d7a2b41aL33-R34)
[[3]](diffhunk://#diff-0d7db8bc789bd7b13874d319644d59a93cd442baa4da9f9a20854434d7a2b41aL42-R55)
[[4]](diffhunk://#diff-0d7db8bc789bd7b13874d319644d59a93cd442baa4da9f9a20854434d7a2b41aL66-R67)
*
[`src/functions/public/Remove-Context.ps1`](diffhunk://#diff-bb64b18ab5c85e130b2713ff666adbabfafe59137c69eb19b1375d410872fbabL28-R28):
Updated the `Remove-Context` function to use `ID` instead of `Name` for
context identification and made the `ID` parameter mandatory.
[[1]](diffhunk://#diff-bb64b18ab5c85e130b2713ff666adbabfafe59137c69eb19b1375d410872fbabL28-R28)
[[2]](diffhunk://#diff-bb64b18ab5c85e130b2713ff666adbabfafe59137c69eb19b1375d410872fbabL42-R42)

### Configuration Updates:
*
[`src/functions/private/Initialize-ContextVault.ps1`](diffhunk://#diff-d31521bac781babd9ce5e82b77fafc5b88a2ad5d69ab8e742bbddd806288b7b2L58-R58):
Set the `VaultName` in the script configuration after initializing and
registering the vault.
[[1]](diffhunk://#diff-d31521bac781babd9ce5e82b77fafc5b88a2ad5d69ab8e742bbddd806288b7b2L58-R58)
[[2]](diffhunk://#diff-d31521bac781babd9ce5e82b77fafc5b88a2ad5d69ab8e742bbddd806288b7b2R70)
*
[`src/functions/private/JsonToObject/ConvertFrom-ContextJson.ps1`](diffhunk://#diff-4b52115c2ce1a1414e44a292fcdff8a38e5e514f5033088475db741661eef168L27-R28):
Added a default value for the `JsonString` parameter.
*
[`src/functions/private/ObjectToJSON/Convert-ContextObjectToHashtableRecursive.ps1`](diffhunk://#diff-ab6ec0a9eedb26e791432c5f00a17e9d6dfba36b39bc9197ed9e97305afe1700L28-R29):
Added a default value for the `Object` parameter.
*
[`src/functions/private/ObjectToJSON/ConvertTo-ContextJson.ps1`](diffhunk://#diff-cce317dfa255618a55a670b2c9a0a0d36faedb455ecf71db9079b127aec77a88L30-R31):
Added a default value for the `Context` parameter.

### Mistake in versioning

Before this PR there was two commits to main that created version
`4.0.4` and `4.0.5` which was a mistake. Both of these were breaking
changes, so they have been reverted and instead gathered into one major
updated, covered in this PR.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas