🩹 [Patch]: Add parameterless class constructor + docs update (#67)
## Description

This pull request includes updates to the `README.md` file and a minor
change to the `Context.ps1` class. The changes in `README.md` focus on
refactoring the module and user configuration instructions to use a more
consistent and simplified approach. Additionally, a constructor was
added to the `Context` class in `Context.ps1`.

Updates to `README.md`:

* Simplified module configuration instructions to use `Get-Context` and
`Set-Context` for modifying settings, instead of individual
`Set-ContextSetting` commands.
* Refactored user configuration instructions to follow a similar
pattern, using `Get-Context` to retrieve the context object, modify it,
and then use `Set-Context` to store the changes.

Update to `Context.ps1`:

* Added a parameterless constructor to the `Context` class to allow for
the creation of context objects without specifying an ID initially.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas